+++ src/test/java/com/dirt/api/usecase/service/TransactionServiceTest.java
@@ -15,6 +15,7 @@ import com.dirt.api.domain.exception.AccountNotExistException;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 
+import java.math.BigDecimal;
 import java.sql.Timestamp;
 import java.util.Optional;
 
@@ -36,8 +37,8 @@ class TransactionServiceTest {
     private static final char ACCOUNT_TYPE = 'F';
     private static final String DES_PAGAMENTO = "Payment Ze Manga";
     private static final String TRANSACTION_IP = "192.168.0.1";
-    private static final Double TRANSACTION_AMOUNT = 100.0;
-    private static final Double TRANSACTION_TAX = 3.0;
+    private static final BigDecimal TRANSACTION_AMOUNT = BigDecimal.valueOf(100.0);
+    private static final BigDecimal TRANSACTION_TAX = BigDecimal.valueOf(3.0);
     private static final String TRANSACTION_CAPTURE_METHOD = "123456";
     private static final CaptureMethodEnum CAPTURE_METHOD = CaptureMethodEnum.APP;
     private static final TransactionTypeEnum TRANSACTION_TYPE = TransactionTypeEnum.PIX;
